+++ Core/Source/ColumnWorkers/BuildableDef/GunStatColumnWorker.cs
@@ -7,7 +7,7 @@ public sealed class GunStatColumnWorker(StatColumnDef columnDef) : StatColumnWor
         // TODO: What about Thing?
         if (@object.Def is Verse.ThingDef { building.turretGunDef: Verse.ThingDef turretGunDef })
         {
-            return base.MakeCell(new DefBasedObject(turretGunDef));
+            return base.MakeCell(new DefBasedObject(turretGunDef, quality: @object.Quality));
         }
 
         return base.MakeCell(@object);